Stop validating instrument unit
