+++ api/all/src/test/java/io/opentelemetry/api/internal/ApiUsageLoggerTest.java
@@ -8,13 +8,15 @@ package io.opentelemetry.api.internal;
 import static java.util.logging.Level.WARNING;
 
 import io.github.netmikey.logunit.api.LogCapturer;
+import io.opentelemetry.internal.testing.slf4j.SuppressLogger;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
 
+@SuppressLogger(ApiUsageLogger.class)
 class ApiUsageLoggerTest {
 
   @RegisterExtension
-  LogCapturer apiUsageLogs = LogCapturer.create().captureForLogger(ApiUsageLogger.LOGGER_NAME);
+  LogCapturer apiUsageLogs = LogCapturer.create().captureForLogger(ApiUsageLogger.class.getName());
 
   @Test
   void log() {
