+++ api/all/src/test/java/io/opentelemetry/api/metrics/DefaultMeterTest.java
@@ -6,20 +6,15 @@
 package io.opentelemetry.api.metrics;
 
 import static io.opentelemetry.api.common.AttributeKey.stringKey;
-import static io.opentelemetry.api.internal.ApiUsageLogger.LOGGER_NAME;
 
-import io.github.netmikey.logunit.api.LogCapturer;
 import io.opentelemetry.api.common.Attributes;
 import io.opentelemetry.context.Context;
 import io.opentelemetry.internal.testing.slf4j.SuppressLogger;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.RegisterExtension;
 
-@SuppressLogger(loggerName = LOGGER_NAME)
+@SuppressLogger()
 public class DefaultMeterTest {
   private static final Meter METER = DefaultMeter.getInstance();
-
-  @RegisterExtension LogCapturer apiUsageLogs = LogCapturer.create().captureForLogger(LOGGER_NAME);
 
   @Test
   void noopLongCounter_doesNotThrow() {
